+++ extensions/youtube/src/main/java/app/revanced/extension/youtube/patches/HidePlayerOverlayButtonsPatch.java
@@ -1,5 +1,7 @@
 package app.revanced.extension.youtube.patches;
 
+import static app.revanced.extension.shared.Utils.getResourceIdentifierOrThrow;
+
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
@@ -37,11 +39,11 @@ public final class HidePlayerOverlayButtonsPatch {
     private static final boolean HIDE_PLAYER_PREVIOUS_NEXT_BUTTONS_ENABLED
             = Settings.HIDE_PLAYER_PREVIOUS_NEXT_BUTTONS.get();
 
-    private static final int PLAYER_CONTROL_PREVIOUS_BUTTON_TOUCH_AREA_ID =
-            Utils.getResourceIdentifier("player_control_previous_button_touch_area", "id");
+    private static final int PLAYER_CONTROL_PREVIOUS_BUTTON_TOUCH_AREA_ID = getResourceIdentifierOrThrow(
+            "player_control_previous_button_touch_area", "id");
 
-    private static final int PLAYER_CONTROL_NEXT_BUTTON_TOUCH_AREA_ID =
-            Utils.getResourceIdentifier("player_control_next_button_touch_area", "id");
+    private static final int PLAYER_CONTROL_NEXT_BUTTON_TOUCH_AREA_ID = getResourceIdentifierOrThrow(
+            "player_control_next_button_touch_area", "id");
 
     /**
      * Injection point.