+++ patches/src/main/kotlin/app/revanced/patches/music/layout/navigationbar/Fingerprints.kt
@@ -1,0 +1,36 @@
+package app.revanced.patches.music.layout.navigationbar
+
+import app.revanced.patcher.fingerprint
+import app.revanced.util.containsLiteralInstruction
+import app.revanced.util.getReference
+import app.revanced.util.indexOfFirstInstruction
+import com.android.tools.smali.dexlib2.AccessFlags
+import com.android.tools.smali.dexlib2.Opcode
+import com.android.tools.smali.dexlib2.iface.Method
+import com.android.tools.smali.dexlib2.iface.reference.MethodReference
+
+internal val tabLayoutTextFingerprint = fingerprint {
+    accessFlags(AccessFlags.PUBLIC, AccessFlags.FINAL)
+    returns("V")
+    parameters("L")
+    opcodes(
+        Opcode.IGET,
+        Opcode.INVOKE_STATIC,
+        Opcode.MOVE_RESULT_OBJECT,
+        Opcode.IF_NEZ,
+        Opcode.SGET_OBJECT,
+        Opcode.INVOKE_INTERFACE,
+        Opcode.MOVE_RESULT
+    )
+    strings("FEmusic_search")
+    custom { method, _ ->
+        method.containsLiteralInstruction(text1)
+                && indexOfGetVisibilityInstruction(method) >= 0
+    }
+}
+
+internal fun indexOfGetVisibilityInstruction(method: Method) =
+    method.indexOfFirstInstruction {
+        opcode == Opcode.INVOKE_VIRTUAL &&
+                getReference<MethodReference>()?.name == "getVisibility"
+    }