+++ patches/src/main/kotlin/app/revanced/patches/youtube/interaction/swipecontrols/SwipeControlsPatch.kt
@@ -46,10 +46,10 @@ private val swipeControlsResourcePatch = resourcePatch {
             ListPreference("revanced_swipe_overlay_style"),
             TextPreference("revanced_swipe_overlay_background_opacity", inputType = InputType.NUMBER),
             TextPreference("revanced_swipe_overlay_progress_brightness_color",
-                tag = "app.revanced.extension.shared.settings.preference.ColorPickerPreference",
+                tag = "app.revanced.extension.shared.settings.preference.ColorPickerWithOpacitySliderPreference",
                 inputType = InputType.TEXT_CAP_CHARACTERS),
             TextPreference("revanced_swipe_overlay_progress_volume_color",
-                tag = "app.revanced.extension.shared.settings.preference.ColorPickerPreference",
+                tag = "app.revanced.extension.shared.settings.preference.ColorPickerWithOpacitySliderPreference",
                 inputType = InputType.TEXT_CAP_CHARACTERS),
             TextPreference("revanced_swipe_text_overlay_size", inputType = InputType.NUMBER),
             TextPreference("revanced_swipe_overlay_timeout", inputType = InputType.NUMBER),
@@ -90,10 +90,9 @@ val swipeControlsPatch = bytecodePatch(
         "com.google.android.youtube"(
             "19.34.42",
             "19.43.41",
-            "19.47.53",
             "20.07.39",
-            "20.12.46",
             "20.13.41",
+            "20.14.43",
         )
     )