+++ patches/src/main/kotlin/app/revanced/patches/youtube/layout/player/fullscreen/OpenVideosFullscreenPatch.kt
@@ -25,10 +25,9 @@ val openVideosFullscreenPatch = bytecodePatch(
 
     compatibleWith(
         "com.google.android.youtube"(
-            "19.47.53",
             "20.07.39",
-            "20.12.46",
             "20.13.41",
+            "20.14.43",
         )
     )