+++ patches/src/main/kotlin/app/revanced/patches/youtube/misc/spoof/SpoofVideoStreamsPatch.kt
@@ -24,10 +24,9 @@ val spoofVideoStreamsPatch = spoofVideoStreamsPatch(
             "com.google.android.youtube"(
                 "19.34.42",
                 "19.43.41",
-                "19.47.53",
                 "20.07.39",
-                "20.12.46",
                 "20.13.41",
+                "20.14.43",
             )
         )
 
@@ -59,7 +58,8 @@ val spoofVideoStreamsPatch = spoofVideoStreamsPatch(
                     ListPreference("revanced_spoof_video_streams_client_type"),
                     NonInteractivePreference(
                         // Requires a key and title but the actual text is chosen at runtime.
-                        key = "revanced_spoof_video_streams_about_android",
+                        key = "revanced_spoof_video_streams_about",
+                        summaryKey = null,
                         tag = "app.revanced.extension.youtube.settings.preference.SpoofStreamingDataSideEffectsPreference"
                     ),
                     ListPreference(